# openapi/ontology.yaml
openapi: 3.1.0
info:
  title: Ontology OpenAPI
  description: Ontology OpenAPI Specification
  version: 0.0.1
servers:
  - url: http://localhost:4242
    description: Local development server
...
paths:
...
  /v1/ontologies/{ontologyId}/explorer/home:
    get:
      tags:
        - Explorer
      summary: Get Explorer Home
      operationId: getExplorerHome
      responses:
        '200':
          description: Get Explorer Home response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExplorerHome'
    parameters:
      - $ref: '#/components/parameters/OntologyId'
...
  /v1/ontologies/{ontologyId}/interfaces/{interfaceId}/object-sets/query:
    post:
      tags:
        - Explorer
      summary: Interface Query
      operationId: interfaceQuery
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterfaceQueryRequest'
      responses:
        '200':
          description: Interface Query response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterfaceQueryPage'
    parameters:
      - name: interfaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
...
  /v1/ontologies/{ontologyId}/object-sets/aggregate:
    post:
      tags:
        - Explorer
      summary: Aggregate
      operationId: aggregate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AggregateRequest'
        required: true
      responses:
        '200':
          description: Aggregate response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregateResponse'
    parameters:
      - $ref: '#/components/parameters/OntologyId'
...
  /v1/ontologies/{ontologyId}/object-sets/query:
    post:
      tags:
        - Explorer
      summary: Query Object Set
      operationId: queryObjectSet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObjectSetRequest'
        required: true
      responses:
        '200':
          description: Query Object Set response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectSetPage'
    parameters:
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/object-sets/search-around:
    post:
      tags:
        - Explorer
      summary: Search Around
      operationId: searchAround
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchAroundRequest'
        required: true
      responses:
        '200':
          description: Search Around response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkPage'
    parameters:
      - $ref: '#/components/parameters/OntologyId'
...
  /v1/ontologies/{ontologyId}/explorer/object-types/{objectTypeId}/objects/{objectId}:
    get:
      tags:
        - Explorer
      summary: Get Object Detail
      operationId: getObjectDetail
      responses:
        '200':
          description: Get Object Detail response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectDetail'
    parameters:
      - name: objectId
        in: path
        required: true
        schema:
          type: string
      - name: objectTypeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/explorer/object-types/{objectTypeId}/objects/{objectId}/activity:
    get:
      tags:
        - Explorer
      summary: List Object Activity
      operationId: listObjectActivity
      responses:
        '200':
          description: List Object Activity response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityItem'
    parameters:
      - name: objectId
        in: path
        required: true
        schema:
          type: string
      - name: objectTypeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/explorer/object-types/{objectTypeId}/objects/{objectId}/capabilities:
    get:
      tags:
        - Explorer
      summary: Get Object Capabilities
      operationId: getObjectCapabilities
      responses:
        '200':
          description: Get Object Capabilities response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'
    parameters:
      - name: objectId
        in: path
        required: true
        schema:
          type: string
      - name: objectTypeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
...
  /v1/ontologies/{ontologyId}/explorer/object-types/{objectTypeId}/objects/{objectId}/provenance:
    get:
      tags:
        - Explorer
      summary: Get Object Provenance
      operationId: getObjectProvenance
      responses:
        '200':
          description: Get Object Provenance response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvenanceView'
    parameters:
      - name: objectId
        in: path
        required: true
        schema:
          type: string
      - name: objectTypeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
...
  /v1/ontologies/{ontologyId}/search/objects:
    post:
      tags:
        - Explorer
      summary: Search Objects
      operationId: searchObjects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        '200':
          description: Search Objects response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
    parameters:
      - $ref: '#/components/parameters/OntologyId'
...
components:
  schemas:
...
    ExplorerHome:
      type: object
      properties:
        objectTypes:
          type: array
          items:
            $ref: '#/components/schemas/ObjectTypeDefinition'
        objectCounts:
          type: object
          additionalProperties:
            type: integer
            format: int64
        searchStatus:
          type: string
        indexUpdatedAt:
          type: string
          format: date-time
...
    InterfaceQueryRequest:
      type: object
      properties:
        pageSize:
          type: integer
          format: int32
    InterfaceQueryPage:
      type: object
      properties:
        interfaceId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/InterfaceObject'
        truncated:
          type: boolean
...
    AggregateRequest:
      type: object
      properties:
        measurePropertyId:
          type: string
          format: uuid
        divisorPropertyId:
          type: string
          format: uuid
        query:
          $ref: '#/components/schemas/ObjectSetRequest'
        dimensionPropertyIds:
          type: array
          items:
            type: string
            format: uuid
        aggregation:
          type: string
    AggregateResponse:
      type: object
      properties:
        objectTypeId:
          type: string
          format: uuid
        queryFingerprint:
          type: string
        buckets:
          type: array
          items:
            $ref: '#/components/schemas/AggregationBucket'
        calculatedAt:
          type: string
          format: date-time
...
    ObjectSetRequest:
      type: object
      properties:
        objectTypeId:
          type: string
          format: uuid
        where:
          type: object
          additionalProperties: {}
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortClause'
        pageSize:
          type: integer
          format: int32
        cursor:
          type: string
        columns:
          type: array
          items:
            type: string
            format: uuid
    ObjectSetPage:
      type: object
      properties:
        objectTypeId:
          type: string
          format: uuid
        properties:
          type: array
          items:
            $ref: '#/components/schemas/PropertyDefinition'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ObjectSummary'
        objectTypeName:
          type: string
        visibleCount:
          type: integer
          format: int64
        countLowerBound:
          type: boolean
        nextCursor:
          type: string
        queryFingerprint:
          type: string
        indexUpdatedAt:
          type: string
          format: date-time
    SearchAroundRequest:
      type: object
      properties:
        objectTypeId:
          type: string
          format: uuid
        objectId:
          type: string
        linkTypeIds:
          type: array
          items:
            type: string
            format: uuid
        pageSize:
          type: integer
          format: int32
    LinkPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ObjectLink'
        nextCursor:
          type: string
        visibleCount:
          type: integer
          format: int64
...
    ObjectDetail:
      type: object
      properties:
        objectId:
          type: string
        title:
          type: string
        properties:
          type: object
          additionalProperties: {}
        objectType:
          $ref: '#/components/schemas/ObjectTypeDefinition'
        redactedFields:
          type: array
          items:
            type: string
            format: uuid
        quality:
          type: string
        updatedAt:
          type: string
          format: date-time
    ActivityItem:
      type: object
      properties:
        correlationId:
          type: string
        kind:
          type: string
        status:
          type: string
        summary:
          type: string
        actor:
          type: string
        occurredAt:
          type: string
          format: date-time
    CapabilityResponse:
      type: object
      properties:
        actions:
          type: array
          items:
            $ref: '#/components/schemas/Capability'
        functions:
          type: array
          items:
            $ref: '#/components/schemas/Capability'
        openTo:
          type: array
          items:
            type: string
...
    ProvenanceView:
      type: object
      properties:
        objectId:
          type: string
        primaryPipeline:
          type: string
        projectionStatus:
          type: string
        sourceAsset:
          type: string
        indexStatus:
          type: string
        fieldLineage:
          type: array
          items:
            type: object
            additionalProperties: {}
    ResourceId:
      type: string
      description: Caller-defined stable ID used in API paths and references
      example: employee
      pattern: ^[A-Za-z][A-Za-z0-9_-]{0,159}$
...
    SearchRequest:
      type: object
      properties:
        query:
          type: string
        mode:
          type: string
        tab:
          type: string
        size:
          type: integer
          format: int32
    SearchResponse:
      type: object
      properties:
        objects:
          type: array
          items:
            $ref: '#/components/schemas/ObjectSummary'
        objectTypes:
          type: array
          items:
            $ref: '#/components/schemas/ObjectTypeDefinition'
        visibleObjectCount:
          type: integer
          format: int64
        indexUpdatedAt:
          type: string
          format: date-time
...
    AggregationBucket:
      type: object
      properties:
        value: {}
        count:
          type: integer
          format: int64
        metric:
          type: number
          format: double
...
    Capability:
      type: object
      properties:
        id:
          type: string
          format: uuid
        apiName:
          type: string
        displayName:
          type: string
        kind:
          type: string
        executable:
          type: boolean
        previewRequired:
          type: boolean
...
    InterfaceObject:
      type: object
      properties:
        objectId:
          type: string
        objectTypeId:
          type: string
          format: uuid
        title:
          type: string
        objectTypeApiName:
          type: string
        slots:
          type: object
          additionalProperties: {}
...
    ObjectLink:
      type: object
      properties:
        relationId:
          type: string
        linkTypeId:
          type: string
          format: uuid
        targetObjectId:
          type: string
        targetObjectTypeId:
          type: string
          format: uuid
        linkTypeName:
          type: string
        direction:
          type: string
        targetTitle:
          type: string
        edgeProperties:
          type: object
          additionalProperties: {}
    ObjectSummary:
      type: object
      properties:
        objectId:
          type: string
        objectTypeId:
          type: string
          format: uuid
        title:
          type: string
        properties:
          type: object
          additionalProperties: {}
        objectTypeApiName:
          type: string
        redactedFields:
          type: array
          items:
            type: string
            format: uuid
        quality:
          type: string
        updatedAt:
          type: string
          format: date-time
    ObjectTypeDefinition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        apiName:
          type: string
        physicalKey:
          type: string
        displayName:
          type: string
        properties:
          type: array
          items:
            $ref: '#/components/schemas/PropertyDefinition'
        maturity:
          type: string
...
    PropertyDefinition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        apiName:
          type: string
        physicalKey:
          type: string
        displayName:
          type: string
        valueType:
          type: string
        primaryKey:
          type: boolean
        titleProperty:
          type: boolean
        searchable:
          type: boolean
        filterable:
          type: boolean
        sortable:
          type: boolean
        sensitive:
          type: boolean
...
    SortClause:
      type: object
      properties:
        propertyId:
          type: string
          format: uuid
        direction:
          type: string
...
  parameters:
...
    OntologyId:
      name: ontologyId
      in: path
      description: Caller-defined stable Ontology ID
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
...